fix int percent

+++ application/views/consent/v_result_admin_detail_from_1.php
@@ -288,7 +288,7 @@ th
                                     </td>
                                     <!-- แสดงเปอร์เซ็นคะแนนรวมทั้งหมด -->
                                     <td align='center'>
-                                   <?php echo   $total_weight/$total*100 ; ?>%
+                                   <?php echo   (int)($total_weight/$total*100) ; ?>%
                                     </td>
                                 </tr>
                             </tbody>
